pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "litlogger"
authors = [
    {name = "Lightning-AI", email = "thomas@lightning.ai"},
]
description = "The Lightning Logger of the Lightning.AI Platform"
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.10"
dynamic = ["version"]
keywords = ["deep learning", "machine learning", "pytorch", "AI"]
classifiers = [
    "Programming Language :: Python :: 3",
]

dependencies = [
    "lightning-sdk >=2025.12.17",
    "litmodels >=0.1.8",
    "protobuf",
    "psutil",
    "rich",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit >= 4.3.0",
]
tests = [
    "coverage",
    "lightning>=2.5.6",
    "pytest>=8.0.0",
    "pytest-timeout>=2.4.0",
    "mock",
    "pytest-mock",
    "pytest-sugar",
    "lightning>=2.0.0",
    "torchvision"
]

[tool.setuptools.dynamic]
version = {attr = "litlogger.__version__"}

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*", "examples*"]

[tool.black]
# https://github.com/psf/black
line-length = 120
exclude = "(.eggs|.git|.hg|.mypy_cache|.venv|_build|buck-out|build|dist)"


[tool.ruff]
target-version = "py310"
line-length = 120
# Exclude a variety of commonly ignored directories.
exclude = [
    ".eggs",
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    "__pypackages__",
    "_build",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",
    "W", # see: https://pypi.org/project/pycodestyle
    "F", # see: https://pypi.org/project/pyflakes
    "I", #see: https://pypi.org/project/isort/
    "D", # see: https://pypi.org/project/pydocstyle
    "N", # see: https://pypi.org/project/pep8-naming
    "UP", # see: pyupgrade
]
extend-select = [
    "A",    # see: https://pypi.org/project/flake8-builtins
    "B",    # see: https://pypi.org/project/flake8-bugbear
    "C4",   # see: https://pypi.org/project/flake8-comprehensions
    "PT",   # see: https://pypi.org/project/flake8-pytest-style
    "RET",  # see: https://pypi.org/project/flake8-return
    "SIM",  # see: https://pypi.org/project/flake8-simplify
    "YTT",  # see: https://pypi.org/project/flake8-2020
    "ANN",  # see: https://pypi.org/project/flake8-annotations
    "TID",  # see: https://pypi.org/project/flake8-tidy-imports/
    "T10",  # see: https://pypi.org/project/flake8-debugger
    "Q",    # see: https://pypi.org/project/flake8-quotes
    "RUF",  # Ruff-specific rules
    "ISC",  # see: https://pypi.org/project/flake8-implicit-str-concat
    "PIE",  # see: https://pypi.org/project/flake8-pie
    "PLE",  # see: https://pypi.org/project/pylint/
    "PYI",  # see: https://pypi.org/project/flake8-pyi/
]
ignore = [
    "D101",   # Missing docstring in public function
    "D102",   # Missing docstring in public function
    "D103",   # Missing docstring in public function
    "E731",   # Do not assign a lambda expression, use a def
    "D100",   # todo: Missing docstring in public module
    "D104",   # todo: Missing docstring in public package
    "D107",   # Missing docstring in `__init__`
    "ANN101", # Missing type annotation for `self` in method
    "UP006",  # Use list instead of List
    "UP035",  # typing.List is deprecated
]
unfixable = ["F401"]

[tool.ruff.lint.per-file-ignores]
#"setup.py" = ["ANN202", "ANN401"]
"src/**" = ["ANN401"]
"examples/**" = [
  "ANN", # Type annotations
  "N",
]
"tests/**" = [
  "ANN", # Type annotations
  "D102", # Missing docstring in public method
  "PT023", # pytest-incorrect-mark-parentheses-style
  "S101", # assert detected
  "E402", #module level import not at top of file
]

[tool.ruff.lint.pydocstyle]
# Use Google-style docstrings.
convention = "google"

#[tool.ruff.pycodestyle]
#ignore-overlong-task-comments = true

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.pytest]
mock_use_standalone_module = "true"
mock_traceback_monkeypatch = "false"
norecursedirs = [
    ".git",
    ".github",
    "dist",
    "build",
    "docs",
]
addopts = [
  "--strict-markers",
  "--doctest-modules",
  "--color=yes",
  "--disable-pytest-warnings",
  "--ignore=legacy/checkpoints",
]
markers = [
    "project_permission_test: marker for project permission tests",
]
filterwarnings = [
  "error::FutureWarning",
  "error::pytest.PytestUnhandledThreadExceptionWarning",
]
xfail_strict = true
junit_duration_report = "call"


[tool.coverage.run]
source = ["lightning_sdk"]
omit = [
    "lightning_sdk/lightning_cloud/*",
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
  "pass",
  "typing.overload",
]
